add api geo gouv address searcher + rearrange files

+++ app/src/main/java/fr/julocorp/jenisassistant/domain/common/Geolocation.kt
@@ -1,6 +1,6 @@
 package fr.julocorp.jenisassistant.domain.common
 
 data class Geolocation(
-    val latitude: Float,
-    val longitude: Float
+    val latitude: Double,
+    val longitude: Double
 )
